pyproject.toml
[tool.poetry]
name = "pjecz-tauro-flask"
version = "0.1.0"
description = "Sistema de Turnos"
authors = ["Guillermo Valdes <guillermo@movimientolibre.com>"]
license = "AGPL3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
click = "^8.1.8"
cryptography = "^44.0.2"
email-validator = "^2.2.0"
flask = "^3.1.0"
flask-login = "^0.6.3"
flask-moment = "^1.0.6"
flask-sqlalchemy = "^3.1.1"
flask-socketio = "^5.5.1"
flask-wtf = "^1.2.2"
gunicorn = "^23.0.0"
hashids = "^1.3.1"
jinja2 = "^3.1.6"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.10"
python = ">=3.11,<4.0"
pydantic = "^2.11.3"
pydantic-settings = "^2.9.1"
pytz = "^2025.2"
sqlalchemy = "^2.0.40"
sqlalchemy-utils = "^0.41.2"
unidecode = "^1.3.8"
uvicorn = "^0.34.2"
werkzeug = "^3.1.3"
flask-restful = "^0.3.10"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
faker = "^37.1.0"
isort = "^6.0.1"
pylint = "^3.3.6"
pylint-sqlalchemy = "^0.3.0"
pytest = "^8.3.5"
requests = "^2.32.3"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 128

[tool.isort]
line_length = 128
profile = "black"
